feat: add interface Film & User
